+++ Documentation/CHANGELOG.md
@@ -58,6 +58,10 @@ Can read pre-2.0.0 logs, but not write them.
 
 - BaseRecorder: Record only when RecordIsReady
 
+### Changed
+
+- Full code cleanup
+
 ## [2.0.0-beta.5] - 2026-09-01
 
 ### Added